+++ packages/tsl-rules-of-react/src/index.ts
@@ -1,0 +1,1 @@
+export { rulesOfReact } from "./rules/rules-of-react";